Allows the selectbox to display selected values with index '0'
Prior to this change the selectbox wasn't able to display selected values (reference-mode:id) with index '0' (as number, not string). An example would be enum properties which are bound to the selectbox by id. Although it's possible to select an enum value with id '0' and save it (assuming we can save the selected value to a database), reloading it from aforementioned database prevents it from beeing displayed in the selectbox, 'undefined' (or nothing) is displayed instead. This change allows for the selectbox to display enum values with index '0' (or values with index '0' in general).